Fixed various issues.
